Add comments about lang setting
The default locale in Pandoc is set to `en-US`, same as the lang settings in Books.jl. To avoid #302, it may be better to switch to other options like `en-GB`. However, this may also have some side effects such as changes in date formats. As a temporary workaround, we can aware the users of this setting.